@@ -1,8 +1,13 @@
 <script lang="ts">
 	import { page } from '$app/stores';
-	export let href: string;
+	interface Props {
+		href: string;
+		children?: import('svelte').Snippet;
+	}
+
+	let { href, children }: Props = $props();
 </script>
 
 <a {href} class={$page.url.pathname == href ? 'text-black dark:text-white' : 'text-gray-500'}
-	><slot /></a
+	>{@render children?.()}</a
 >
